requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
sentence-transformers>=2.2.0
faiss-cpu>=1.9.0
langchain-google-vertexai>=1.0.0
langgraph>=0.0.50
langchain>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
redis>=5.0.0
python-multipart>=0.0.6